Update page free-space calculation implementation in layers/page.c

+++ layers/page.c
@@ -5,12 +5,11 @@
 #include "../headers/page.h"
 #include "../headers/slot.h"
 
-
-int get_free_space(Page * page) {
+size_t get_free_space(Page * page) {
      PageHeader header = page->header;
 
      int last_slot_ends_at = sizeof(PageHeader) + header.slot_count*sizeof(Slot);
-     int free_space = page->header.free_space_pointer - last_slot_ends_at;
+     size_t free_space = page->header.free_space_pointer - last_slot_ends_at;
 
      return  free_space;
 }
